\name{aRpsDCA-package}
\alias{aRpsDCA-package}
\alias{aRpsDCA}

\docType{package}

\title{
Arps Decline Curve Analysis in R
}

\description{
Functions for Arps decline-curve analysis. Includes exponential, hyperbolic, harmonic, and hyperbolic-to-exponential models.
}

\details{
\tabular{ll}{
Package: \tab aRpsDCA\cr
Version: \tab 0.1.0\cr
Date: \tab 2014-03-03\cr
Authors@R: \tab c(person("Derrick", "Turk", role = c("aut", "cre", "cph"),
                     email = "dwt@terminusdatascience.com"))\cr
Depends: \tab \cr
License: \tab file LICENSE\cr
License_is_FOSS: \tab yes\cr
License_restricts_use: \tab yes\cr
URL: \tab https://github.com/derrickturk/aRpsDCA\cr
BugReports: \tab https://github.com/derrickturk/aRpsDCA/issues\cr
Packaged: \tab 2014-03-04 20:18:28 UTC; derrick.turk\cr
Built: \tab R 3.0.2; ; 2014-03-04 20:18:35 UTC; windows\cr
}

Index:
\preformatted{
as.effective            Arps decline conversion from nominal to
                        effective
as.nominal              Arps decline conversion from effective to
                        nominal
exponential             Arps exponential declines
harmonic                Arps harmonic declines
hyp2exp                 Arps hyperbolic-to-exponential declines
hyperbolic              Arps hyperbolic declines
rescale.by.time         Time unit conversion for DCA
}
}

\author{
Derrick Turk [aut, cre, cph]

Maintainer: Derrick Turk <dwt@terminusdatascience.com>
}

\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }

\examples{
~~ simple examples of the most important functions ~~
}
