\name{curtailed}
\alias{curtailed}
\alias{curtail}
\alias{curtailed.q}
\alias{curtailed.Np}
\alias{curtailed.D}
\alias{arps.q.curtailed}
\alias{arps.Np.curtailed}
\alias{arps.D.curtailed}

\title{Arps decline curves with initial curtailment}

\description{
    Create decline curve objects and compute rates, cumulative production, and nominal declines for Arps decline curves with an initial period of curtailment to constant rate.
}

\usage{
    curtailed.q(t.curtail, decline, t)
    curtailed.Np(t.curtail, decline, t)
    curtailed.D(t.curtail, decline, t)
    curtail(decline, t.curtail)
}
